+++ src/main/java/net/geforcemods/securitycraft/blocks/ScannerTrapDoorBlock.java
@@ -1,0 +1,35 @@
+package net.geforcemods.securitycraft.blocks;
+
+import net.geforcemods.securitycraft.blockentities.ScannerTrapdoorBlockEntity;
+import net.geforcemods.securitycraft.blocks.reinforced.BaseIronTrapDoorBlock;
+import net.geforcemods.securitycraft.misc.OwnershipEvent;
+import net.minecraft.block.Block;
+import net.minecraft.block.material.Material;
+import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.EntityLivingBase;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.item.ItemStack;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import net.minecraftforge.common.MinecraftForge;
+
+public class ScannerTrapDoorBlock extends BaseIronTrapDoorBlock {
+	public ScannerTrapDoorBlock(Material material) {
+		super(material);
+	}
+
+	@Override
+	public void onBlockPlacedBy(World world, BlockPos pos, IBlockState state, EntityLivingBase placer, ItemStack stack) {
+		if (placer instanceof EntityPlayer)
+			MinecraftForge.EVENT_BUS.post(new OwnershipEvent(world, pos, (EntityPlayer) placer));
+	}
+
+	@Override
+	public void neighborChanged(IBlockState state, World world, BlockPos pos, Block block, BlockPos neighbor) {}
+
+	@Override
+	public TileEntity createNewTileEntity(World world, int meta) {
+		return new ScannerTrapdoorBlockEntity();
+	}
+}